pyproject.toml
[project]
name = "llm-security-proxy"
version = "0.1.0"
description = "FastAPI reverse proxy with security detectors for LLM requests"
authors = [{name = "Open Source"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=1.10,<2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N"]
ignore = []

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
